requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
cssselect==1.2.0
docopt==0.6.2
eutils==0.6.0
frozenlist==1.4.1
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
habanero==1.2.6
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
humanfriendly==10.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-core==0.3.6
langchain-text-splitters==0.3.0
langsmith==0.1.129
Levenshtein==0.26.0
lxml==5.3.0
lxml_html_clean==0.2.2
metapub==0.5.12
multidict==6.1.0
mutagen==1.47.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dotenv==1.0.1
python-Levenshtein==0.26.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.0
requests==2.32.3
rsa==4.9
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tabulate==0.9.0
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
websockets==13.1
yarl==1.13.1
yt-dlp==2024.9.27
